# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MO2men Coffee Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MenuItem {
  id: number;
  /** Item name in Arabic */
  name: string;
  description?: string;
  category: string;
  price: number;
  imageUrl?: string;
  available: boolean;
  /** Estimated preparation minutes */
  prepMinutes: number;
  createdAt: string;
}

export interface MenuItemInput {
  name: string;
  description?: string;
  category: string;
  price: number;
  imageUrl?: string;
  available?: boolean;
  prepMinutes?: number;
}

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  category?: string;
  price?: number;
  imageUrl?: string;
  available?: boolean;
  prepMinutes?: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface OrderItem {
  menuItemId: number;
  name: string;
  quantity: number;
  unitPrice: number;
  notes?: string;
}

export interface OrderItemInput {
  menuItemId: number;
  quantity: number;
  notes?: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  served: "served",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  tableNumber: number;
  customerName?: string;
  status: OrderStatus;
  items: OrderItem[];
  totalPrice: number;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export interface OrderInput {
  tableNumber: number;
  customerName?: string;
  items: OrderItemInput[];
  notes?: string;
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  served: "served",
  cancelled: "cancelled",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface TodayStats {
  ordersCount: number;
  revenue: number;
  activeOrders: number;
  averagePrepMinutes: number;
  servedOrders: number;
}

export interface PopularItem {
  menuItemId: number;
  name: string;
  category: string;
  quantitySold: number;
}

export interface QueueSummary {
  pending: number;
  preparing: number;
  ready: number;
  /** How long the oldest non-served order has been waiting */
  oldestWaitingSeconds: number;
}

export type ListMenuParams = {
  category?: string;
  availableOnly?: boolean;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  /**
   * Only orders that are not served or cancelled
   */
  activeOnly?: boolean;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  preparing: "preparing",
  ready: "ready",
  served: "served",
  cancelled: "cancelled",
} as const;

export type GetPopularItemsParams = {
  limit?: number;
};