# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MO2men Coffee Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  /** Item name in Arabic */
  name: string;
  description?: string;
  category: string;
  price: number;
  imageUrl?: string;
  available: boolean;
  /** Estimated preparation minutes */
  prepMinutes: number;
  createdAt: Date;
}