# lib/api-zod/src/generated/types/menuItemUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MO2men Coffee Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  category?: string;
  price?: number;
  imageUrl?: string;
  available?: boolean;
  prepMinutes?: number;
}