# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MO2men Coffee Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  tableNumber: number;
  customerName?: string;
  status: OrderStatus;
  items: OrderItem[];
  totalPrice: number;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
}