# lib/api-zod/src/generated/types/popularItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MO2men Coffee Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface PopularItem {
  menuItemId: number;
  name: string;
  category: string;
  quantitySold: number;
}