# lib/api-zod/src/generated/types/queueSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MO2men Coffee Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface QueueSummary {
  pending: number;
  preparing: number;
  ready: number;
  /** How long the oldest non-served order has been waiting */
  oldestWaitingSeconds: number;
}